+++ app/Helpers/SystemBase.php
@@ -7,7 +7,6 @@ enum SystemBase: string
     case DRE = 'DRE';
     case OpScan = 'OpScan';
     case Computer = 'PC/Laptop';
-    case Hardware = 'Hardware';
     case Tablet = 'Tablet';
     case CustomHardware = 'Custom Hardware';
     case Other = 'Other';
