+++ tests/Unit/DispositionTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Unit;
 
 use App\Models\Disposition;
 use App\Models\InventoryUnit;
